refactor(cli): split pure analysis functions from helpers.ts into analysis.ts
Move stateless computation and formatting functions to a new
cli/src/analysis.ts module, keeping I/O-heavy bootstrap code in
helpers.ts.

Functions moved to analysis.ts:
- SessionData and DailyEntry interfaces
- effectiveTokens() -- pure token selection helper
- getEditorSourceFromPath() -- pure path classification
- runWithConcurrency() -- generic bounded-concurrency utility
- createEmptyPeriodStats() -- factory for PeriodStats objects
- aggregateIntoPeriod() -- accumulates session fractions into a period
- createEmptyUsageAnalysisPeriod() -- factory for UsageAnalysisPeriod
- buildChartPayload() -- transforms daily stats into chart JSON
- fmt() / formatTokens() -- pure number formatters
- CHART_COLORS constant

helpers.ts retains:
- Ecosystem adapter bootstrap (getEcosystems, createSessionDiscovery)
- Session file I/O (processSessionFile, statSessionFile)
- Async aggregation pipelines (calculateDetailedStats, calculateDailyStats,
  calculateUsageAnalysisStats, buildCustomizationMatrix)
- Discovery and diagnostics (discoverSessionFiles, getDiagnosticPaths)

All existing public exports are preserved via re-exports in helpers.ts,
so no consumer changes are needed. npm run build passes.
